--- app/constants/search.js
@@ -1,3 +1,0 @@
-export default Object.freeze({
-    DEFAULT_LIMIT: 10
-});
