--- app/cors/allowedOrigins.js
@@ -1,7 +1,0 @@
-const allowedOrigins = [
-    'https://www.yoursite.com',
-    'http://127.0.0.1:4000',
-    'http://localhost:4000'
-];
-
-export { allowedOrigins };
