--- app/cors/corsOptions.js
@@ -1,15 +1,0 @@
-import { StatusCodes } from "http-status-codes";
-import { allowedOrigins } from "./allowedOrigins.js";
-
-const corsOptions = {
-    origin: (origin, callback) => {
-        if (allowedOrigins.indexOf(origin) !== -1 || !origin) {
-            callback(null, true)
-        } else {
-            callback(new Error('Not allowed by CORS'));
-        }
-    },
-    optionsSuccessStatus: StatusCodes.OK
-}
-
-export { corsOptions }
