--- app/db/connect.js
@@ -1,7 +1,0 @@
-import mongoose from "mongoose";
-
-const connectDB = (url) => {
-    return mongoose.connect(url);
-}
-
-export { connectDB }
