--- app/middleware/credentials.js
@@ -1,11 +1,0 @@
-import { allowedOrigins } from "../cors/allowedOrigins.js";
-
-const credentials = (req, res, next) => {
-    const origin = req.headers.origin;
-    if (allowedOrigins.includes(origin)) {
-        res.header('Access-Control-Allow-Credentials', true);
-    }
-    next();
-}
-
-export { credentials }
