--- app/models/user.js
@@ -1,66 +1,0 @@
-import mongoose, { Schema as Schema, model } from "mongoose";
-import jwt from "jsonwebtoken";
-
-const UserSchema = new mongoose.Schema({
-  name: {
-    type: String,
-    required: [true, "Name is required"],
-    maxlength: 50,
-    minlength: 2,
-  },
-  email: {
-    type: String,
-    required: [true, "Email is required"],
-    match: [
-      /^(([^<>()[\]\\.,;:\s@"]+(\.[^<>()[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/,
-      "Please provide a valid email",
-    ],
-    unique: true,
-  },
-  emailVerified: {
-    type: Boolean,
-    default: false,
-  },
-  phone: {
-    type: String,
-    required: false,
-    match: [/^\+[1-9]{1}[0-9]{3,14}$/, "Please provide a valid phone number"],
-    unique: false,
-  },
-  phoneVerified: {
-    type: Boolean,
-    default: false,
-  },
-  sharedSecret: {
-    type: String,
-    required: false,
-  },
-  refreshToken: {
-    type: String,
-    required: false,
-  },
-});
-
-UserSchema.methods.createAccessToken = function () {
-  return jwt.sign(
-    { userId: this._id, name: this.name, email: this.email },
-    process.env.ACCESS_TOKEN_SECRET,
-    {
-      expiresIn: process.env.ACCESS_TOKEN_EXPIRATION,
-    }
-  );
-};
-
-UserSchema.methods.createRefreshToken = function () {
-  return jwt.sign(
-    { userId: this._id, name: this.name, email: this.email },
-    process.env.REFRESH_TOKEN_SECRET,
-    {
-      expiresIn: process.env.REFRESH_TOKEN_EXPIRATION,
-    }
-  );
-};
-
-const User = mongoose.model("User", UserSchema);
-
-export { User };
