--- app/package.json
@@ -1,32 +1,0 @@
-{
-  "name": "recipes-backend-v2",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "type": "module",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "debug": "nodemon index.js",
-    "start": "node index.js"
-  },
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "@types/bcryptjs": "^2.4.6",
-    "bcryptjs": "^2.4.3",
-    "cookie-parser": "^1.4.6",
-    "cors": "^2.8.5",
-    "deepl-node": "^1.13.0",
-    "dotenv": "^16.4.5",
-    "express": "^4.19.2",
-    "express-async-errors": "^3.1.1",
-    "http-status-codes": "^2.3.0",
-    "jsonwebtoken": "^9.0.2",
-    "moment": "^2.30.1",
-    "mongoose": "^8.3.1",
-    "mustache": "^4.2.0",
-    "nodemailer": "^6.9.13",
-    "nodemon": "^3.1.0",
-    "otp-generator": "^4.0.1"
-  }
-}
