+++ package.json.old
@@ -1,0 +1,11 @@
+{
+  "devDependencies": {
+    "@types/dotenv": "^8.2.0",
+    "@types/express": "^4.17.21",
+    "@types/jsonwebtoken": "^9.0.6",
+    "@types/mustache": "^4.2.5",
+    "@types/nodemailer": "^6.4.15",
+    "@types/otp-generator": "^4.0.2",
+    "typescript": "^5.4.5"
+  }
+}
