+++ src/app/constants/lang.ts
@@ -1,0 +1,3 @@
+export default Object.freeze({
+  DEFAULT_LANG: "fr",
+});
