+++ src/app/constants/search.ts
@@ -1,0 +1,3 @@
+export default Object.freeze({
+  DEFAULT_LIMIT: 10,
+});
