+++ src/app/cors/allowedOrigins.ts
@@ -1,0 +1,7 @@
+const allowedOrigins: string[] = [
+  "https://www.yoursite.com",
+  "http://127.0.0.1:4000",
+  "http://localhost:4000",
+];
+
+export { allowedOrigins };
