+++ src/app/db/connect.ts
@@ -1,0 +1,7 @@
+import mongoose from "mongoose";
+
+const connectDB = (url: string) => {
+  return mongoose.connect(url);
+};
+
+export { connectDB };
