+++ src/app/middleware/notFound.ts
@@ -1,0 +1,8 @@
+import { StatusCodes } from "http-status-codes";
+import { Request, Response } from "express";
+
+const notFound = (req: Request, res: Response) => {
+  res.status(StatusCodes.NOT_FOUND).send("Route not found");
+};
+
+export { notFound };
