+++ src/app/types/custom.ts
@@ -1,0 +1,5 @@
+export type AuthenticatedUser = {
+  userId: string;
+  name: string;
+  email: string;
+};
