+++ src/app/types/express/index.d.ts
@@ -1,0 +1,13 @@
+import { AuthenticatedUser } from "../custom";
+
+// to make the file a module and avoid the TypeScript error
+export {};
+
+declare global {
+  namespace Express {
+    export interface Request {
+      //language?: Language;
+      user?: AuthenticatedUser;
+    }
+  }
+}
